refactor: resolve circular dependency between pricing-fetcher and macro modules
- Move ModelPricing type and ModelPricingSchema from pricing-fetcher.ts to types.internal.ts
- Update imports in pricing-fetcher.ts, macro.internal.ts, and pricing-fetcher.test.ts
- Fix ESLint import ordering issues in test file
- Eliminate circular dependency where pricing-fetcher imported from macro.internal via macro import,
  while macro.internal imported types from pricing-fetcher

This change improves module architecture by centralizing shared types in the types module,
preventing the circular import chain that could cause build or runtime issues.
